+++ Block/Form/AdditionalInfo.php
@@ -3,6 +3,8 @@
  * Copyright © Alekseon sp. z o.o.
  * http://www.alekseon.com/
  */
+declare(strict_types=1);
+
 namespace Alekseon\WidgetForms\Block\Form;
 
 use Magento\Framework\View\Element\Template;