+++ CHANGELOG.md
@@ -10,7 +10,50 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 ### Removed
 
-## [Unreleased]
+## [102.4.2] - 2024-05-24
+### Changed
+- Translate labels (https://github.com/Alekseon/AlekseonEav/issues/45)
+
+## [102.4.1] - 2024-02-26
+### Fixed
+- fix ajax error: "complete is not a function"
+
+## [102.4.0] - 2023-12-03
+### Changed
+- refactoring - removed UI component on frontend
+
+## [102.3.0] - 2023-06-21
+### Changed
+- frontend inputs moved from Alekseon_WidgetForms to Alekseon_CustonFormsFrontend
+### Added
+- form wrapper class
+
+## [102.2.7] - 2023-05-02
+### Changed
+- changes for tests in github actions
+
+## [102.2.6] - 2023-05-01
+### Added
+- added email validation for newsletter email input in admin
+### Fixed
+- code quality fixes (https://github.com/Alekseon/magento2-widget-forms/issues/17)
+
+## [102.2.3] - 2023-03-28
+### Added
+- field containers css classes (for hyva)
+
+## [102.2.2] - 2023-03-12
+### Fixed
+- fix for cache issue
+- removed gap and fix label displaye in image input template
+
+## [102.2.1] - 2023-03-12
+### Added
+- "*" for required fields
+- cache tags
+### Fixed
+- select default value on radio button input
+- stop opening first tab after submit if there is only one tab
 
 ## [102.2.0] - 2023-03-04
 ### Added